refactor: separate user group
